'job-app' was updated to the repository

+++ job-app/src/main/java/com/example/jobapp/company/impl/CompanyServiceImpl.java
@@ -43,8 +43,11 @@ public class CompanyServiceImpl implements CompanyService {
 
     @Override
     public Company deleteCompanyById(int companyId) {
-        Company company = getCompanyById(companyId);
-        companyRepository.deleteById(companyId);
-        return company;
+        if(companyRepository.existsById(companyId)){
+            Company company = getCompanyById(companyId);
+            companyRepository.deleteById(companyId);
+            return company;
+        }
+        return null;
     }
 }
